requirements.txt
Flask
pandas
scikit-learn
tensorflow-cpu
gunicorn
numpy